# Disable babel, improve test speed, use Chrome headless...