--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,4 @@ npm-debug.log
 sample/dist
 *.swp
 .idea/
+package-lock.json