--- a/js/config.js
+++ b/js/config.js
@@ -13,7 +13,7 @@ defaultsDeep(systematic, {
     public_path: '/',
     enable_filename_hashing: true,
     enable_babel: true,
-    es_libs: [],
+    no_parse: [],
   },
   serve: {
     // FIXME(vperron): This configuration comes in double with the one from the makefile.