--- a/js/webpack_get_defaults.js
+++ b/js/webpack_get_defaults.js
@@ -134,7 +134,6 @@ function buildBabelPresets (profile) {
 
 function getBabelPlugins () {
   const plugins = [
-    'transform-strict-mode',
     ['transform-runtime', {polyfill: false}],
     'transform-class-properties',
   ]