--- a/js/webpack_get_defaults.js
+++ b/js/webpack_get_defaults.js
@@ -150,6 +150,7 @@ module.exports = function (basePath) {
 
   const PATHS = {
     src: path.join(basePath, config.build.src_dir),
+    es_libs: config.build.es_libs.map((es_lib) => path.join(basePath, 'node_modules', es_lib)),
     dist: path.join(basePath, config.build.output_dir),
   }
 
@@ -265,8 +266,7 @@ module.exports = function (basePath) {
         {
           test: /\.(js|jsx)/,
           use: ['cache-loader', ...jsLoaders],
-          exclude: /(node_modules|bower_components)/,
-          include: [PATHS.src],
+          include: [PATHS.src, ...PATHS.es_libs],
         },
         {
           test: /\.vue$/,