--- a/package.json
+++ b/package.json
@@ -83,7 +83,6 @@
     "postcss-loader": "2.0.6",
     "puppeteer": "^0.12.0",
     "sass-loader": "6.0.6",
-    "source-map-loader": "^0.2.2",
     "standard": "10.0.3",
     "style-loader": "0.18.2",
     "uglify-js": "3.1.3",