--- a/package.json
+++ b/package.json
@@ -55,6 +55,7 @@
     "eslint-plugin-html": "3.2.2",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "1.1.4",
+    "hard-source-webpack-plugin": "^0.4.11",
     "html-loader": "0.5.1",
     "html-webpack-plugin": "2.30.1",
     "http-server": "0.10.0",