--- a/package.json
+++ b/package.json
@@ -42,8 +42,6 @@
     "babel-plugin-angularjs-annotate": "0.8.2",
     "babel-plugin-transform-class-properties": "6.24.1",
     "babel-plugin-transform-runtime": "6.23.0",
-    "babel-plugin-transform-strict-mode": "6.24.1",
-    "babel-polyfill": "6.26.0",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-es2016": "6.24.1",
     "babel-preset-es2017": "6.24.1",